add ability to run setup
