+++ noxopt/_config.py
@@ -1,0 +1,4 @@
+import os
+
+
+NOXOPT_TESTING = bool(int(os.getenv("NOXOPT_TESTING", 0)))
