+++ pyproject.toml
@@ -19,7 +19,7 @@ classifiers = [
 dependencies = [
     "nox",
     'importlib-metadata; python_version<"3.8"',
-    'typing_extensions; python_version<"3.11"',
+    'typing_extensions; python_version<"3.10"',
 ]
 
 [project.urls]
