+++ requirements/check-tests.txt
@@ -1,0 +1,2 @@
+pytest
+coverage
