+++ requirements/check-types.txt
@@ -1,0 +1,1 @@
+mypy
