+++ tests/__init__.py
@@ -1,0 +1,4 @@
+import os
+
+
+os.environ["NOXOPT_TESTING"] = "1"
